move icon files and add props
