+++ src/components/social-media-component/SocialMediaComponentNoBackground.tsx
@@ -1,6 +1,6 @@
 import styles from './assets/social-media-component-no-background.module.scss';
-import FacebookIcon from './assets/FacebookIcon';
-import InstagramIcon from './assets/InstagramIcon';
+import FacebookIcon from '../../icons/FacebookIcon';
+import InstagramIcon from '../../icons/InstagramIcon';
 import React from 'react';
 
 export default function SocialMediaComponentNoBackground() {
