+++ src/icons/FacebookIcon.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
+import { IconProps } from '../Contract';
 
-export default function FacebookIcon() {
+export default function FacebookIcon(props: IconProps) {
     return (
         <a href="https://www.facebook.com/pwrinspace">
             <svg width={33} height={33} viewBox="0 0 33 33" fill="none" xmlns="http://www.w3.org/2000/svg">
